+++ src/components/modals/ReferenceModal.vue
@@ -58,7 +58,7 @@
             class="icon"
             :style="{
               backgroundImage: `url(${require(
-                '../../assets/icons/' + jinx.first.id + '.png',
+                '../../assets/icons/' + jinx.first.id + '.webp',
               )})`,
             }"
           ></span>
@@ -66,7 +66,7 @@
             class="icon"
             :style="{
               backgroundImage: `url(${require(
-                '../../assets/icons/' + jinx.second.id + '.png',
+                '../../assets/icons/' + jinx.second.id + '.webp',
               )})`,
             }"
           ></span>
@@ -152,7 +152,7 @@ export default {
 
       return require('../../assets/icons/' +
         (role.imageAlt || role.id) +
-        '.png');
+        '.webp');
     },
     ...mapMutations(["toggleModal"]),
   },